Feat: Docs MCP overhaul, new tools and updated architecture to keep it dynamic
