--- juspay_docs_mcp/mcpdoc/__init__.py
@@ -1,3 +1,0 @@
-from juspay_docs_mcp.mcpdoc._version import __version__
-
-__all__ = ["__version__"]
