--- juspay_docs_mcp/mcpdoc/_version.py
@@ -1,7 +1,0 @@
-from importlib import metadata
-
-try:
-    __version__ = metadata.version(__package__)
-except metadata.PackageNotFoundError:
-    # Case where package metadata is not available.
-    __version__ = ""
